\name{Adjusted RV correlation between two sets of variables}
\alias{arv}
\title{
Adjusted  RV correlation between two sets of variables
}

\description{
Adjusted RV correlation between two sets of variables.
}

\usage{
arv(y, x)
}

\arguments{
\item{y}{
A numerical matrix.
}
\item{x}{
A numerical matrix.
}
}

\details{
The adjusted RV correlation coefficient is computed.
}

\value{
The value of the adjusted RV coefficient.
}

\references{
Mordant G. and Segers J. (2022). Measuring dependence between random vectors via optimal transport. Journal of Multivariate Analysis, 189: 104912.
}

\author{
Michail Tsagris

R implementation and documentation: Michail Tsagris \email{mtsagris@uoc.gr}.
}

%\note{
%%  ~~further notes~~
%}

\seealso{
\code{ \link{mrv}, \link{rv}, \link{drv}, \link{sq.correl}, \link{dcor}
}
}

\examples{
arv( as.matrix(iris[, 1:2]), as.matrix(iris[, 3:4]) )
}


