\name{Modified RV correlation between two sets of variables}
\alias{mrv}
\title{
Modified RV correlation between two sets of variables
}

\description{
Modified RV correlation between two sets of variables.
}

\usage{
mrv(y, x)
}

\arguments{
\item{y}{
A numerical matrix.
}
\item{x}{
A numerical matrix.
}
}

\details{
The modified RV correlation coefficient
}

\value{
The value of the modified RV coefficient.
}

\references{
Smilde A. K., Kiers H. A., Bijlsma S., Rubingh C. M. and Van Erk M. J. (2009). Matrix correlations for high-dimensional data: the modified RV-coefficient. Bioinformatics, 25(3): 401--405.
}

\author{
Michail Tsagris

R implementation and documentation: Michail Tsagris \email{mtsagris@uoc.gr}.
}

%\note{
%%  ~~further notes~~
%}

\seealso{
\code{ \link{rv}, \link{arv}, \link{drv}, \link{sq.correl}, \link{dcor}
}
}

\examples{
mrv( as.matrix(iris[, 1:2]), as.matrix(iris[, 3:4]) )
}


